requirements.txt

streamlit>=1.28.0
pandas>=2.0.0
requests>=2.31.0
PyYAML>=6.0
plotly>=5.17.0
altair>=5.0.0
openpyxl>=3.1.0
schedule>=1.2.0
python-dateutil>=2.8.2
numpy>=1.24.0
datetime
pathlib
